user authentication for creating and deleting events

+++ backend/controllers/calendarEvent.js
@@ -2,14 +2,16 @@ const CalendarEvent = require("../models/calendarEvent");
 
 
 exports.createCalendarEvent = (req, res, next) => {
-    console.log("test" + req.params.roomId);
+    console.log(req.body);
     const calendarEvent = new CalendarEvent({
         title: req.body.title,
         start: req.body.start,
         end: req.body.end,
         allDay: req.body.allDay,
-        roomId: req.params.roomId
+        roomId: req.params.roomId,
+        creator: req.body.userId
     });
+
     calendarEvent.save().then(createdCalendarEvent => {
         res.status(201).json({
             message: "Event sucessfully added!",
@@ -19,6 +21,7 @@ exports.createCalendarEvent = (req, res, next) => {
             }
         });
     }).catch(error => {
+        console.log(error);
         res.status(500).json({
             message: "Event creation failed!"
         })
@@ -28,9 +31,7 @@ exports.createCalendarEvent = (req, res, next) => {
 
 //holt alle Events
 exports.getCalendarEvents = (req, res, next) => {
-    console.log("get" + req.params.roomId);
     CalendarEvent.find({ 'roomId': req.params.roomId }).then(calendarEvents => {
-        console.log(calendarEvents);
         if (calendarEvents) {
             res.status(200).json({ message: "Events fetched successfully", calendarEvents: calendarEvents });
         } else {
@@ -60,7 +61,7 @@ exports.getCalendarEvent = (req, res, next) => {
 };
 
 exports.deleteCalendarEvent = (req, res, next) => {
-    CalendarEvent.deleteOne({ _id: req.params.id }).then(result => {
+    CalendarEvent.deleteOne({ _id: req.params.id, creator: req.params.userId }).then(result => {
         console.log(result);
         if (result.deletedCount > 0) {
             res.status(200).json({ message: "Event sucessfully deleted!" });
@@ -81,10 +82,10 @@ exports.updateOne = (req, res, next) => {
         title: req.body.title,
         start: req.body.start,
         end: req.body.end,
-        allDay: req.body.allDay
+        allDay: req.body.allDay,
+        creator: req.body.creator
     });
-    console.log(calendarEvent);
-    CalendarEvent.updateOne({ _id: req.params.id/*, creator: req.userData.userId*/ }, calendarEvent).then(result => {
+    CalendarEvent.updateOne({ _id: req.params.id, creator: req.userData.userId }, calendarEvent).then(result => {
         if (result.matchedCount > 0) {
             res.status(200).json({ message: "Event sucessfully edited!" });
         }
